requirements.txt
fastapi
uvicorn
openai
googletrans==4.0.0rc1
python-dotenv
